% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trnobj_doc.R
\name{as_list}
\alias{as_list}
\title{Convert an TRN object to a list}
\usage{
TRNobj$as_list()
}
\value{
A list
}
\description{
This method extracts all fields of an TRN object and places their stored values 
into the fields of an R list, with list field names matching the TRN object field names.
}
