% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trnobj_doc.R
\name{CADJ_active_verts}
\alias{CADJ_active_verts}
\title{Return the CADJ active vertices}
\usage{
TRNobj$CADJ_active_verts()
}
\value{
a list of vertex IDs that are connected by any \code{CADJ_active} edge.
}
\description{
Active CADJ vertices participate in active CADJ edges; 
Thus if there are vertices in the graph with no active CADJ edges connecting them 
they are deemed inactive.
}
