% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trnobj_doc.R
\name{calc_DMPrune_LambdaPath}
\alias{calc_DMPrune_LambdaPath}
\title{Calculate the DM-Prune Lambda Path}
\usage{
TRNobj$calc_DMPrune_LambdaPath(priorADJ)
}
\arguments{
\item{priorADJ}{an adjacency matrix (nrows = ncols = \code{nV}) whose (i,j) entries contain the Dirichlet prior CADJ values.}
}
\description{
The DM-Prune Lambda path, at each prune step \eqn{t}, is the (normalized) Dirichlet-Multinomial likelihood of CADJ edges 
after pruning edges whose \code{DMP_mu_rank} is \strong{strictly less than} \eqn{t}.
}
