% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trnobj_doc.R
\name{calc_TopoMeasures}
\alias{calc_TopoMeasures}
\title{Calculate Topology Preserving Measures of the TRN}
\usage{
TRNobj$calc_TopoMeasures

\insertRef{BauerPawelzikGeisel1992}{TopoRNet}
\insertRef{VillmannDerHerrmannMartinetz1997}{TopoRNet}
\insertRef{ZhangMerenyi2006}{TopoRNet}
}
\description{
Sets the Topology Preserving Measures of both the CADJ and CONN graphs (Topographic Product and several Topographic Functions). 
See \code{?Topographic_Product} and \code{?Topographic_Functions} for more details.
}
