% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trnobj_doc.R
\name{CONN_active_verts}
\alias{CONN_active_verts}
\title{Return the CONN active vertices}
\usage{
TRNobj$CONN_active_verts()
}
\value{
a list of vertex IDs that are connected by any \code{CONN_active} edge.
}
\description{
Active CONN vertices participate in active CONN edges; 
Thus if there are vertices in the graph with no active CONN edges connecting them 
they are deemed inactive.
}
