% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dDirMult}
\alias{dDirMult}
\title{Density of Dirichlet-Multinomial Distribution}
\usage{
dDirMult(x, prior, log_form = TRUE, normalize = TRUE)
}
\arguments{
\item{x}{the count vector at which to evaluate the DM density}

\item{prior}{the prior count vector, must have \code{length = length(x)}}

\item{log_form}{whether to return the log-density, default = TRUE}

\item{normalize}{whether to normalize the calculated density by \code{sum(x)}. 
Default = TRUE.}
}
\value{
The evaluated DM density (a number)
}
\description{
Computes the density of a Dirichlet-Multinomial (DM) distribution, at a given 
count vector \code{x}, with a user-specified prior count vector \code{prior}.
}
