% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trnobj_doc.R
\name{DMPrune_CADJ_step}
\alias{DMPrune_CADJ_step}
\title{Prune CADJ at a DM-Prune step}
\usage{
TRNobj$DMPrune_CADJ_step(min_step)
}
\arguments{
\item{min_step}{}
}
\description{
Prune the CADJ graph by removing edges in \code{CADJ_EL} whose \code{DMP_mu_rank} is \strong{strictly less than} a given minimum prune step.
}
