% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trnobj_doc.R
\name{get_CADJ}
\alias{get_CADJ}
\title{Return the CADJ adjacency matrix}
\usage{
TRNobj$get_CADJ()
}
\value{
an adjacency matrix (nrows = ncols = \code{nV}) whose (i,j) entries are the \code{CADJ} edge weights.
}
\description{
This method returns the CADJ adjacency matrix of the TRN (as set during \code{initialize_TRN}) whose \code{CADJ_active} flag = 1.
}
