% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trnobj_doc.R
\name{get_CADJvis_colors}
\alias{get_CADJvis_colors}
\title{Get the CADJvis edge colors}
\usage{
TRNobj$get_CADJvis_colors()
}
\value{
A vector containing the CADJvis (hex) color of each edge in \code{CADJ_EL}.
}
\description{
The edge colorings (red, blue, green, yellow, grayscale) are assigned in increasing order of an edge's \code{CADJ_lrank}.
}
\references{
\insertRef{TasdemirMerenyi2009}{TopoRNet}
}
