% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trnobj_doc.R
\name{get_CADJvis_widths}
\alias{get_CADJvis_widths}
\title{Get the CADJvis edge widths}
\usage{
TRNobj$get_CADJvis_widths()
}
\value{
A vector containing the CADJvis widths of each edge in \code{CADJ_EL}.
}
\description{
The edge widths are assigned in reverse of each edge's \code{CADJ_grank}. That is, if a CADJ graph has unique global ranks 
of 1,2,3,4,5, then edges with those global ranks will have widths = 5,4,3,2,1.
}
\references{
\insertRef{TasdemirMerenyi2009}{TopoRNet}
}
