% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trnobj_doc.R
\name{get_CONN}
\alias{get_CONN}
\title{Return the CONN adjacency matrix}
\usage{
TRNobj$get_CONN()
}
\value{
an adjacency matrix (nrows = ncols = \code{nV}) whose (i,j) entries are the \code{CONN} edge weights.
}
\description{
This method returns the CONN adjacency matrix of the TRN (as set during \code{initialize_TRN}) whose \code{CONN_active} flag = 1.
}
