% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trnobj_doc.R
\name{get_CONNvis_colors}
\alias{get_CONNvis_colors}
\title{Get the CONNvis edge colors}
\usage{
TRNobj$get_CONNvis_colors()
}
\value{
A vector containing the CONNvis (hex) color of each edge in \code{CONN_EL}.
}
\description{
The edge colorings (red, blue, green, yellow, grayscale) are assigned in increasing order of an edge's \code{CONN_lrank}.
}
\references{
\insertRef{TasdemirMerenyi2009}{TopoRNet}
}
