% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trnobj_doc.R
\name{get_CONNvis_stats}
\alias{get_CONNvis_stats}
\title{Get the CONNvis summary statistics.}
\usage{
TRNobj$get_CONNvis_stats(group)
}
\arguments{
\item{group}{a string identifying which summary statistics to return. Can be one of 'lrank', 'grank', or 'length'.}
}
\value{
A data frame with summary statistics for each group in its rows, and columns: 
\itemize{
\item \code{lrank} (or \code{grank} or \code{length} as applicable), identifying the grouping for each set of summary stats
\item \code{count} the number of CONN edges in the group 
\item \code{pct} the proportion of the total number of CONN edges in each group 
\item \code{cumpct} the cumulative proportion of CADH edges in each group (as ordered by the grouping variable)
\item \code{mean} the average of CONN edge weights in each group 
\item \code{sd} the standard deviation of CONN edge weights in each group 
\item \code{q0} the 0.00 quantile (minimum) of CONN edge weights in each group 
\item \code{q25} the 0.25 quantile (Q1) of CONN edge weights in each group 
\item \code{q50} the 0.50 quantile (median) of CONN edge weights in each group 
\item \code{q75} the 0.75 quantile (Q3) of CONN edge weights in each group
\item \code{q100} the 1.00 quantile (maximum) of CONN edge weights in each group
}
}
\description{
The CONNvis summary statistics are computed for each unique grouping found in \code{CONN_lrank}, \code{CONN_grank}, and \code{CONN_fwdflen}.
}
\references{
\insertRef{TasdemirMerenyi2009}{TopoRNet}
}
