% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trnobj_doc.R
\name{get_CONNvis_widths}
\alias{get_CONNvis_widths}
\title{Get the CONNvis edge widths}
\usage{
TRNobj$get_CONNvis_widths()
}
\value{
A vector containing the CONNvis widths of each edge in \code{CONN_EL}.
}
\description{
The edge widths are assigned in reverse of each edge's \code{CONN_grank}. That is, if a CONN graph has unique global ranks 
of 1,2,3,4,5, then edges with those global ranks will have widths = 5,4,3,2,1.
}
\references{
\insertRef{TasdemirMerenyi2009}{TopoRNet}
}
