% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trnobj_doc.R
\name{get_OTADJ}
\alias{get_OTADJ}
\title{Return the TRN's output space adjacency matrix}
\usage{
TRNobj$get_OTADJ()
}
\value{
an adjacency matrix (nrows = ncols = \code{nV})
}
\description{
This method returns the adjacency matrix of the output space topology of the TRN 
(as set during \code{initialize_TRN}),
}
