% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trnobj_doc.R
\name{initialize_TRN}
\alias{initialize_TRN}
\title{Initialize a TRN object}
\usage{
TRNobj$initialize_TRN(CADJ, OTADJ)
\insertRef{MartinetzSchulten1994}{TopoRNet}
}
\arguments{
\item{CADJ}{The CADJ adjacency matrix of the TRN.}

\item{OTADJ}{The adjacency matrix of the TRN's output topology (e.g., the SOM neuron lattice adjacency). 
Only binary output topology adjacency matrices are supported at this time. 
(\eqn{OTADJ_{ij} = 1} indicate an edge between vertices \code{i} and \code{j}).}
}
\value{
None
}
\description{
Sets the CADJ, CONN, and output topology graphs.
}
\details{
The following fields are computed and stored internally: 
\itemize{
\item \code{nV}, set = nrow(CADJ). 
\item \code{CADJ_EL} and \code{CONN_EL}, the CADJ and CONN edge lists. 
\item \code{CADJ} and \code{CONN}, the CADJ and CONN edge weights. 
\item \code{CADJ_nE} and \code{CONN_nE}, the number of edges in the CADJ and CONN graphs. 
\item \code{CADJ_deg} and \code{CONN_deg}, the degree of each TRN vertex in the CADJ and CONN graphs. 
\item \code{CADJ_wdeg} and \code{CONN_wdeg}, the weighted degree of each TRN vertex in the CADJ and CONN graphs. 
\item \code{OT_EL} and \code{OT_nE} and \code{OT_deg} the edge list, number of edges, and vertex degrees in the TRN's output topology. 
\item \code{OT_max_dist}, the maximum geodesic distance of any two vertices in the TRN's output topology. 
\item \code{CADJ_fwdflen} and \code{CONN_fwdflen}, the forward folding lengths of each CADJ and CONN edge. 
\item \code{CADJ_bwdflen} and \code{CONN_bwdflen}, the backward folding lengths of each output topology edge on the CADJ and CONN graphs. 
\item \code{CADJ_TP_radius} and \code{CONN_TP_radius}, the Topology Preserving Radius of the CADJ and CONN graphs. 
\item \code{CADJ_lrank} and \code{CONN_lrank} The local ranks of each CADJ and CONN edge. 
\item \code{CADJ_grank} and \code{CONN_grank} The global ranks of each CADJ and CONN edge. 
}

Both \code{CADJ} and \code{OTADJ} should be square, and have the same dimensions. 
In addition to the above list of internally computed quantities, \code{initialize_TRN} also computes the 
CADJvis and CONNvis summary statistics (retrievable via \link{get_CADJvis_stats} and \link{get_CONNvis_stats}, respectively), and 
sets all \code{CADJ_active} and \code{CONN_active} flags = 1.
}
