% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trnobj_doc.R
\name{load}
\alias{load}
\title{Load an existing TRN object}
\usage{
TRNobj$load(TRNfile)
}
\arguments{
\item{trnfile}{a string indicating the file path and name of the saved TRN object.}
}
\value{
None, the TRN object is loaded
}
\description{
TRN objects previously written to disk via the \link{save} method can be re-loaded into 
a new R environment with this function. All fields of the internal C++ class will be populated, and 
all methods can be called on the loaded TRN object.
}
\details{
Because the .trn file is in .rds format it can, technically, be loaded directly into an R environment 
as a list via \link[base]{readRDS}. This can be useful for spot checking the contents of a saved TRN object, but 
does not allow use of any of its methods (or visualizations).  The \code{load} method allows for 
proper restoration of a previously saved TRN
}
