% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trnobj_doc.R
\name{new}
\alias{new}
\title{Create an empty TRN object}
\usage{
TRN$new()
}
\value{
an empty TRN object templated for initialization (via \link{initialize_TRN}).
}
\description{
Create an empty TRN object
}
