% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trnobj_doc.R
\name{prune_CADJ_edge_list}
\alias{prune_CADJ_edge_list}
\title{Prune CADJ by an edge list}
\usage{
TRNobj$prune_CADJ_edge_list(edge_list)
}
\arguments{
\item{edge_list}{a 2-column matrix rows define the vertices connected by each CADJ edge to prune}
}
\value{
None
}
\description{
Prune the CADJ graph given a list of edges to remove.
}
\details{
The vertex indices in \code{edge_list} should be 1-based.
}
