% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trnobj_doc.R
\name{prune_CADJ_grank}
\alias{prune_CADJ_grank}
\title{Prune CADJ by a maximum global rank}
\usage{
TRNobj$prune_CADJ_grank(max_grank)
}
\arguments{
\item{max_grank}{maximum \code{CADJ_grank} allowed in the pruned graph. 
Any edges whose \code{CADJ_grank} is > \code{max_grank} will be pruned.}
}
\value{
None
}
\description{
Prune the CADJ graph by removing edges whose \code{CADJ_grank} is \strong{strictly greater than} a given maximum grank.
}
