% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trnobj_doc.R
\name{prune_CADJ_length}
\alias{prune_CADJ_length}
\title{Prune CADJ by a maximum output topology length}
\usage{
TRNobj$prune_CADJ_length(max_fwdflen)
}
\arguments{
\item{max_fwdflen}{maximum \code{CADJ_fwdflen} allowed in the pruned graph. 
Any edges whose \code{CADJ_fwdflen} is > \code{max_fwdflen} will be pruned.}
}
\value{
None
}
\description{
Prune the CADJ graph by removing edges whose \code{CADJ_fwdflen} is \strong{stricly greater than} a given maximim forward folding length.
}
