% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trnobj_doc.R
\name{prune_CADJ_lrank}
\alias{prune_CADJ_lrank}
\title{Prune CADJ by a maximum local rank}
\usage{
TRNobj$prune_CADJ_lrank(max_lrank)
}
\arguments{
\item{max_lrank}{maximum \code{CADJ_lrank} allowed in the pruned graph. 
Any edges whose \code{CADJ_lrank} is > \code{max_lrank} will be pruned.}
}
\value{
None
}
\description{
Prune the CADJ graph by removing edges whose \code{CADJ_lrank} is \strong{strictly greater than} a given maximum lrank.
}
