% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trnobj_doc.R
\name{prune_CADJ_vertex_id}
\alias{prune_CADJ_vertex_id}
\title{Prune CADJ by a set of vertex ids}
\usage{
TRNobj$prune_CADJ_vertex_id(vertex_id)
}
\arguments{
\item{vertex_id}{a vector containing the vertices to prune. 
All connections in \code{CADJ_EL} from or to the vertices in \code{vertex_id} will be pruned.}
}
\value{
None
}
\description{
Prune the CADJ graph by removing edges connecting vertices in a list of given vertex IDs.
}
