% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trnobj_doc.R
\name{prune_CONN_edge_id}
\alias{prune_CONN_edge_id}
\title{Prune CONN by edge ids}
\usage{
TRNobj$prune_CONN_edge_id(edge_id)
}
\arguments{
\item{edge_id}{a vector containing the rows indices of \code{CONN_EL} which contain the edges desired to prune.}
}
\value{
None
}
\description{
Prune the CONN graph given a list of edge IDs to remove.
}
