% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trnobj_doc.R
\name{prune_CONN_grank}
\alias{prune_CONN_grank}
\title{Prune CONN by a maximum global rank}
\usage{
TRNobj$prune_CONN_grank(max_grank)
}
\arguments{
\item{max_grank}{maximum \code{CONN_grank} allowed in the pruned graph. 
Any edges whose \code{CONN_grank} is > \code{max_grank} will be pruned.}
}
\value{
None
}
\description{
Prune the CONN graph by removing edges whose \code{CONN_grank} is \strong{strictly greater than} a given maximum grank.
}
