% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trnobj_doc.R
\name{prune_CONN_length}
\alias{prune_CONN_length}
\title{Prune CONN by a maximum output topology length}
\usage{
TRNobj$prune_CONN_length(max_fwdflen)
}
\arguments{
\item{max_fwdflen}{maximum \code{CONN_fwdflen} allowed in the pruned graph. 
Any edges whose \code{CONN_fwdflen} is > \code{max_fwdflen} will be pruned.}
}
\value{
None
}
\description{
Prune the CONN graph by removing edges whose \code{CONN_fwdflen} is \strong{stricly greater than} a given maximim forward folding length.
}
