% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trnobj_doc.R
\name{prune_CONN_lrank}
\alias{prune_CONN_lrank}
\title{Prune CONN by a maximum local rank}
\usage{
TRNobj$prune_CONN_lrank(max_lrank)
}
\arguments{
\item{max_lrank}{maximum \code{CONN_lrank} allowed in the pruned graph. 
Any edges whose \code{CONN_lrank} is > \code{max_lrank} will be pruned.}
}
\value{
None
}
\description{
Prune the CONN graph by removing edges whose \code{CONN_lrank} is \strong{strictly greater than} a given maximum lrank.
}
