% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trnobj_doc.R
\name{prune_CONN_vertex_id}
\alias{prune_CONN_vertex_id}
\title{Prune CONN by a set of vertex ids}
\usage{
TRNobj$prune_CONN_vertex_id(vertex_id)
}
\arguments{
\item{vertex_id}{a vector containing the vertices to prune. 
All connections in \code{CONN_EL} from or to the vertices in \code{vertex_id} will be pruned.}
}
\value{
None
}
\description{
Prune the CONN graph by removing edges connecting vertices in a list of given vertex IDs.
}
