% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trnobj_doc.R
\name{prune_CONN_wdeg}
\alias{prune_CONN_wdeg}
\title{Prune CONN by a minimum vertex weight}
\usage{
TRNobj$prune_CONN_wdeg(min_weight)
}
\arguments{
\item{min_weight}{minimum vertex weight allowed in the pruned graph. 
Any vertices whose \code{CONN_wdeg} is < \code{min_weight} will be pruned.}
}
\value{
None
}
\description{
Prune the CONN graph by removing edges incident to vertices whose \code{CONN_wdeg} is \strong{stricly less than} a given minimum weighted degree.
}
