% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trnobj_doc.R
\name{restore_CADJ_edges}
\alias{restore_CADJ_edges}
\title{Restore all CADJ edges}
\usage{
TRNobj$restore_CADJ_edges()
}
\value{
None
}
\description{
When pruning of CADJ edges occurs (via any of the \code{prune_CADJ_*} methods), the \code{CADJ_active} flag corresponding 
to the pruned edges is changed from 1 to 0. This method restores all \code{CADJ_active} flags to 1 (reversing any existing pruning).
}
