% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trnobj_doc.R
\name{restore_CONN_edges}
\alias{restore_CONN_edges}
\title{Restore all CONN edges}
\usage{
TRNobj$restore_CONN_edges()
}
\value{
None
}
\description{
When pruning of CONN edges occurs (via any of the \code{prune_CONN_*} methods), the \code{CONN_active} flag corresponding 
to the pruned edges is changed from 1 to 0. This method restores all \code{CONN_active} flags to 1 (reversing any existing pruning).
}
