% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trnobj_doc.R
\name{save}
\alias{save}
\title{Save a TRN object}
\usage{
TRNobj$save(trnfile)
}
\arguments{
\item{trnfile}{a string indicating the file path and name in which to save the TRN object.
This must end in extension ".trn", otherwise an error is returned.}
}
\value{
None, the TRN object is saved to disk
}
\description{
All fields in a TRN object can be saved to disk with this function, which allows them to be re-loaded 
into a new R environment at a later time.
}
\details{
The TRN object is saved to disk as an R list, with each field occupying a corresponding field of the list.  
The file is saved in .rds format (can check its details with \link[base]{infoRDS}).  

Saved TRNs can be re-loaded with \link{load}
}
