% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trnobj_doc.R
\name{set_BootSig}
\alias{set_BootSig}
\title{Set the CADJ significances}
\usage{
TRNobj$set_BootSig(BSADJ)
}
\arguments{
\item{BSADJ}{an adjacency matrix (nrows = ncols = \code{nV}) whose (i,j) entries contain significance values 
of each CADJ edge.}
}
\description{
Set the Bootstrap significance values of each CADJ and CONN edge.
}
