% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trnobj_doc.R
\name{set_parallel}
\alias{set_parallel}
\title{Setter function for parallel computation.}
\usage{
TRNobj$set_parallel(parallel)
}
\arguments{
\item{parallel}{either TRUE or FALSE, as desired}
}
\value{
None, \code{parallel} field is set internally
}
\description{
Sets an internal flag controlling whether computations are performed in parallel.
}
\details{
Parallel computation is supported via RcppParallel, 
see \link[RcppParallel]{setThreadOptions} for details to control threading. 
By default, \code{parallel = TRUE} at TRNobj instantiation.
}
