% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Topographic_Functions}
\alias{Topographic_Functions}
\title{Topographic Functions of a TRN}
\usage{
Topographic_Functions(inputADJ, outputADJ, parallel = TRUE)
}
\arguments{
\item{inputADJ}{the adjacency matrix of vertices in the Input space of the TRN}

\item{outputADJ}{the adjacency matrix of vertices in the Output sapce of the TRN}

\item{parallel, }{whether to compute in parallel. Default = T.}
}
\value{
a data frame with columns: 
\itemize{
\item \code{k} list of folding lengths
\item \code{TF} the TF computed at each folding length 
\item \code{DTF} the DTF computed at each folding length 
\item \code{WDTF} the WDTF computed at each folding length 
}

\insertRef{VillmannDerHerrmannMartinetz1997}{TopoRNet}
\insertRef{ZhangMerenyi2006}{TopoRNet}
}
\description{
The Topograph Functions of a TRN measure topology preservation across a range of "folding" lengths, 
which are defined as the length (geodesic distance) by which edges in one space (input / output) must "fold" in order 
to be represented in the other.  

The functions computed here are the TF (Topographic Function) of Villmann et al. and the 
DTF (Differential Topographic Function ) & WDTF (Weighted DTF) of Zhang et al.
}
