% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Topographic_Product}
\alias{Topographic_Product}
\title{Topographic Product of a TRN}
\usage{
Topographic_Product(input_coords, output_coords, parallel = TRUE)
}
\arguments{
\item{input_coords}{the coordinates of the vertices in the Input space (i.e, R^d) of the TRN}

\item{output_coords}{the coordinates of the vertices in the Output space of the TRN}

\item{parallel, }{whether to compute in parallel. Default = T.}
}
\value{
The Topographic Product (a number)
}
\description{
The Topographic Product of a TRN which is imbued with an output topology.
}
\references{
\insertRef{BauerPawelzikGeisel1992}{TopoRNet}
}
