% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_trn.R
\name{vis_CONN_TopoFunctions}
\alias{vis_CONN_TopoFunctions}
\title{Visualize the Topographic Functions for CONN}
\usage{
vis_CONN_TopoFunctions(TRN)
}
\arguments{
\item{TRN}{a TRN object}
}
\description{
Visualize the Topographic Functions for CONN
}
\details{
Three subplots are returned, corresponding to the Topographic Function, the Discrete Topographic Function, and 
the Weighted Discrete Topographic Function, respectively. 
See \code{?Topographic_Functions} for further details.
}
