% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_trn.R
\name{vis_CONNvis}
\alias{vis_CONNvis}
\title{CONNvis Visualization}
\usage{
vis_CONNvis(
  TRN,
  add = F,
  vertex.xy,
  vertex.pch = 16,
  vertex.cex = 1,
  vertex.col = "black",
  edge.lwd_range = NULL,
  vertex.active = T,
  vertex.subset = NULL
)
}
\arguments{
\item{TRN}{a TRN object}

\item{add}{whether to create a new plotting device (=FALSE, default), or add to an existing one (=TRUE)}

\item{vertex.xy}{a matrix of (x,y) coordinates (nrows = TRN$nV, ncols = 2) defining the plot coordinates 
of the vertices of the TRN. It is assumed that the \code{i-th} row of \code{vertex.xy} gives the coordinates for vertex \code{i} of the graph.}

\item{vertex.pch}{the pch symbol plotted vertices, default = 16.}

\item{vertex.cex}{the cex of plotted vertices, default = 1. Set = 0 to suppress vertex plotting.}

\item{vertex.col}{the color of plotted vertices, default = 'black'}

\item{edge.lwd_range}{the min/max range of the plotted line widths. 
Default = NULL means line widths inherit from the \code{TRN$CONN_grank} of each edge. 
If supplied as something other than NULL, it must be a length=2 vector giving the (lower,upper) bounds of plotted edge widths.}

\item{vertex.active}{whether to restrict plotted vertices to those which have an edge connecting them, default = TRUE.}

\item{vertex.subset}{a vector of vertex indices to restrict the plotting to. Default = NULL, meaning all edges and vertices are plotted. 
If given, any edges connecting vertices in this list (along with the vertices themselves) will not be plotted.}
}
\description{
CONNvis Visualization
}
\details{
Only active CONN edges are plotted (i.e., those with \code{TRN$CONN_active = 1}), 
so that any previously pruned CONN edges are not shown.  If all edges are desired, called \code{TRN$restore_CONN_edges} 
prior to plotting.
}
