% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_trn.R
\name{vis_CONNvis_stats}
\alias{vis_CONNvis_stats}
\title{Visualize CONN Statistics}
\usage{
vis_CONNvis_stats(TRN, group_by = "lrank")
}
\arguments{
\item{TRN}{a TRN object}

\item{group_by}{which grouping to visualize, default = 'lrank'. Can also be 'grank' or 'length'}
}
\description{
The CONNvis statistics will be plotted by each local rank, global rank, or output topology distance, as requested.
}
