% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_dmprune.R
\name{vis_DMPrune_LambdaPath}
\alias{vis_DMPrune_LambdaPath}
\title{Visualize the DM-Prune Lambda Path}
\usage{
vis_DMPrune_LambdaPath(
  TRN,
  min_step = NULL,
  max_step = NULL,
  scale = 1,
  plot.cpts = T
)
}
\arguments{
\item{TRN}{a TRN object}

\item{min_step}{the minimum pruning step to show}

\item{max_step}{the maximum pruning step to show}

\item{scale}{a scale factor controlling the size of plotted lines, 
points, and text labels. Higher values increase the size.}

\item{plot.cpts}{whether to compute and plot the variance changepoints of the 
second-order differenced Lambda path. Default = TRUE.}
}
\value{
None, a plot is produced.
}
\description{
Visualize the DM-Prune Lambda Path
}
