% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_dmprune.R
\name{vis_DMPrune_prior}
\alias{vis_DMPrune_prior}
\title{Visualize the DM-Prune prior}
\usage{
vis_DMPrune_prior(TRN)
}
\arguments{
\item{TRN}{a TRN object}
}
\value{
None, a plot is produced.
}
\description{
Plots the CADJ prior values vs. the CADJ values for comparison.
}
