% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_trn.R
\name{vis_TopoView}
\alias{vis_TopoView}
\title{TopoView Visualization}
\usage{
vis_TopoView(
  ADJ,
  add = F,
  vertex.xy,
  vertex.pch = 16,
  vertex.cex = 1,
  vertex.col = "black",
  edge.col = "darkorange",
  edge.lwd_range = c(1, 5),
  vertex.active = T,
  vertex.subset = NULL
)
}
\arguments{
\item{ADJ}{a (possibly weighted) adjacency matrix of a Topology Representing Network.}

\item{add}{whether to create a new plotting device (=FALSE, default), or add to an existing one (=TRUE)}

\item{vertex.xy}{a matrix of (x,y) coordinates (nrows = TRN$nV, ncols = 2) defining the plot coordinates of the vertices of the TRN. 
It is assumed that the \code{i-th} row of \code{vertex.xy} gives the coordinates for vertex \code{i} of the graph.}

\item{vertex.pch}{the pch symbol plotted vertices, default = 16.}

\item{vertex.cex}{the cex of plotted vertices, default = 1. Set = 0 to suppress vertex plotting.}

\item{vertex.col}{the color of plotted vertices.}

\item{edge.lwd_range}{the min/max range of the plotted line widths, Default = c(1, 5). 
This parameter is only valid if input \code{ADJ} is weighted, in which case the line widths represent 
the edge weights via a linear scaling from \code{(min weight, max weight)} to \code{edge.lwd_range}. 
If \code{ADJ} is unweighted, the plotted edges have width = \code{max(edge.lwd_range)}.}

\item{vertex.active}{whether to restrict plotted vertices to those which have an edge connecting them, default = TRUE.}

\item{vertex.subset}{a vector of vertex indices to restrict the plotting to. Default = NULL, meaning all edges and vertices are plotted. 
If given, any edges connected vertices in this list (along with the vertices themselves) will not be plotted.}

\item{edge.color}{line color of plotted edges, default = "darkorange".}
}
\description{
TopoView Visualization
}
\details{
It is assumed that any 0 value in the input \eqn{ADJ_{ij}} means no edge connects vertices \eqn{i} and \eqn{j}.
}
